requirements.txt
fastapi
uvicorn
playwright
mysql-connector-python
python-dotenv